move multiplication logic to separate file
