+++ ff-macros/src/small_fp/mod.rs
@@ -1,4 +1,5 @@
 mod montgomery_backend;
+mod mul;
 mod utils;
 
 use quote::quote;
